refactor: 패키지 구조 변경
- webserver: TCP 소켓, HTTP 요청 파싱 등 네트워크가 주요 관심사
- servlet: 파싱된 HttpRequest에 대한 정적/동적 자원 처리가 주요 관심사
