--- src/main/java/codesquad/handler/Handler.java
@@ -1,10 +1,0 @@
-package codesquad.handler;
-
-import codesquad.http.HttpRequest;
-import codesquad.http.HttpResponse;
-
-public interface Handler {
-
-    void service(HttpRequest request, HttpResponse response);
-
-}
