+++ src/main/java/codesquad/webserver/http/HttpMethod.java
@@ -1,4 +1,4 @@
-package codesquad.http;
+package codesquad.webserver.http;
 
 public enum HttpMethod {
