+++ src/main/java/codesquad/webserver/http/HttpPath.java
@@ -1,5 +1,6 @@
-package codesquad.http;
+package codesquad.webserver.http;
 
+import codesquad.webserver.SupportFileExtension;
 import java.util.Arrays;
 import java.util.Collections;
 import java.util.Map;
