+++ src/main/java/codesquad/webserver/http/HttpRequestLine.java
@@ -1,4 +1,4 @@
-package codesquad.http;
+package codesquad.webserver.http;
 
 public class HttpRequestLine {
