+++ src/main/java/codesquad/webserver/http/HttpRequestParser.java
@@ -1,4 +1,4 @@
-package codesquad.http;
+package codesquad.webserver.http;
 
 import java.io.BufferedReader;
 import java.io.IOException;
