+++ src/main/java/codesquad/webserver/http/HttpVersion.java
@@ -1,4 +1,4 @@
-package codesquad.http;
+package codesquad.webserver.http;
 
 import java.util.Arrays;
